Major refactor of DS Rolls & type updates
