+++ src/module/data/message/saving-throw.mjs
@@ -1,6 +1,6 @@
 import BaseMessageModel from "./base.mjs";
 
-/** @import { SavingThrowRoll } from "../../rolls/savingThrow.mjs" */
+/** @import { SavingThrowRoll } from "../../rolls/saving-throw.mjs" */
 /** @import DrawSteelActiveEffect from "../../documents/active-effect.mjs"; */
 
 const fields = foundry.data.fields;
