+++ src/module/rolls/_module.mjs
@@ -1,5 +1,5 @@
-export { DSRoll } from "./base.mjs";
-export { DamageRoll } from "./damage.mjs";
-export { PowerRoll } from "./power.mjs";
-export { ProjectRoll } from "./project.mjs";
-export { SavingThrowRoll } from "./savingThrow.mjs";
+export { default as DSRoll } from "./base.mjs";
+export { default as DamageRoll } from "./damage.mjs";
+export { default as PowerRoll } from "./power.mjs";
+export { default as ProjectRoll } from "./project.mjs";
+export { default as SavingThrowRoll } from "./saving-throw.mjs";
