+++ src/module/rolls/saving-throw.mjs
@@ -1,10 +1,10 @@
-import { DSRoll } from "./base.mjs";
+import DSRoll from "./base.mjs";
 import { systemPath } from "../constants.mjs";
 
 /**
  * Usually a flat d10 roll to shake free of a persistent condition
  */
-export class SavingThrowRoll extends DSRoll {
+export default class SavingThrowRoll extends DSRoll {
   /**
    * @param {string} [formula="1d10"]        Default saving throw is a flat 1d10
    * @param {Record<string, any>} [data]     Roll data
@@ -34,22 +34,15 @@ export class SavingThrowRoll extends DSRoll {
     return this.total >= this.successThreshold;
   }
 
-  /**
-   * Helper function to generate render context in use with `static CHAT_TEMPLATE`
-   * @param {object} options
-   * @param {string} [options.flavor]     Flavor text to include
-   * @param {boolean} [options.isPrivate] Is the Roll displayed privately?
-   * @returns An object to be used in `renderTemplate`
-   */
-  async _prepareContext({ flavor, isPrivate }) {
-    return {
-      formula: isPrivate ? "???" : this._formula,
-      flavor: isPrivate ? null : flavor ?? this.options.flavor ?? game.i18n.localize("DRAW_STEEL.Roll.Save.Label"),
-      user: game.user.id,
-      tooltip: isPrivate ? "" : await this.getTooltip(),
-      total: isPrivate ? "?" : Math.round(this.total * 100) / 100,
-      result: this.product ? "critical" : "failure",
-    };
+  /** @inheritdoc */
+  async _prepareChatRenderContext({ flavor, isPrivate = false, ...options } = {}) {
+    const context = await super._prepareChatRenderContext({ flavor, isPrivate, ...options });
+
+    if (!isPrivate) context.flavor ??= game.i18n.localize("DRAW_STEEL.Roll.Save.Label");
+
+    context.result = this.product ? "critical" : "failure";
+
+    return context;
   }
 
   /** @inheritdoc */
